사복검 fx 2타가 나가지 않던 문제 수정

+++ Project_Team_Thief/Assets/Game/Player/Script/Unit/PlayerUnit.cs
@@ -819,7 +819,7 @@ public class PlayerUnit : Unit
         playerInfo.CurHP = _curHp;
         //---
 
-        if (_curHp <= 0)
+        if ((int)_curHp <= 0)
         {
             _curHp = 0;
             playerInfo.CurHP = _curHp;
